+++ packages/react/src/components/Provider/ProviderConsumer.tsx
@@ -2,7 +2,8 @@ import * as PropTypes from 'prop-types'
 import * as React from 'react'
 import { FelaTheme } from 'react-fela'
 
-import { ThemePrepared, ProviderContextPrepared } from '../../themes/types'
+import { ThemePrepared } from '../../themes/types'
+import { ProviderContextPrepared } from '../../types'
 
 export interface ProviderConsumerProps {
   /**
