+++ packages/react/src/lib/createComponent.tsx
@@ -7,8 +7,7 @@ import renderComponent, { RenderResultConfig } from './renderComponent'
 import { AccessibilityActionHandlers } from './accessibility/types'
 import { FocusZone } from './accessibility/FocusZone'
 import { createShorthandFactory } from './factories'
-import { ObjectOf } from '../types'
-import { ProviderContextPrepared } from '../themes/types'
+import { ObjectOf, ProviderContextPrepared } from '../types'
 
 export interface CreateComponentConfig<P> {
   displayName: string
